Polymorphic Select helpers.
Provides a polymorphic_select helper to ActionView and FormBuilder to
make it easier to work with polymorphic select elements.

Makes use of Rails Multiparameter Assignment much like the
datetime_select helpers to pass the information back to the model. As
such it overrides part of the Multiparameter Assignment process to
allow for the assignment of polymorphic associations.